Update CSS and JS for responsive design
Update App.css and home.tsx files to make the site more responsive.
In App.css, the styles were changed so that the height and width are now at 100% instead of 100vh and 100vw.
In home.tsx, the renderer size was changed to fill the whole window instead of just half,
and a resize event listener was added to update the renderer and camera size on window resize.
This creates a more user-friendly, responsive design.